Move bloom filter to separate module
